+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Make `utils` module private as it only contains helper functions for testing. Remove dummy tests.
+
 ## [0.5.8] - 2024-03-25
 
 * Fixed calculating checksums for new files